feat: add password authentication flow (#2023)
This commit adds support for logging in via password using the Auth0
Universal Login screen.

It keeps around the existing magic link support, but relegates it to a
secondary (well, stylistically tertiary) button. Native form submit of
the email field now goes to the password login.

To ease users into it, we added explanatory text advertising the
feature and also directing them to the sign up page, although they can
also sign up from the login page itself.

It also replaces a few `this.state` direct-usages with destructured
assignments as a small bit of inline cleanup.

Co-authored-by: Whitman Schorn <whitman.schorn@gmail.com>